tests/test_requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.25.0
python-dotenv>=0.19.0 